Fold the logical subscription registry back into the stable client
The split was lint collateral, not design. The base client sat at 297
effective lines against a 300 ceiling; the fence's admission calls
pushed it over and it was carved up to fit. With the admission code
gone the subscriptions fit inline again, so the transport returns to
exactly what main ships: stable-logical-rpc-client.ts and its test are
now byte-identical to aac38d698f, and 297 effective lines is back under
the ceiling with no per-file bump.
