+++ mobile/src/transport/stable-logical-rpc-client.ts
@@ -7,7 +7,6 @@ import {
 import { waitForAuthenticated } from './replacement-session-authentication'
 import { projectMobileRpcRequestParams } from './mobile-rpc-request-projection'
 import { LogicalClientConnectionPath } from './logical-client-connection-path'
-import { LogicalSubscriptionRegistry } from './logical-subscription-registry'
 
 export type MobileConnectionPath = 'lan' | 'tailscale' | 'relay'
 
@@ -23,6 +22,15 @@ export function isLogicalClientCutoverError(error: unknown): boolean {
     error instanceof LogicalClientCutoverError ||
     (error instanceof Error && error.message === 'RPC interrupted by connection migration')
   )
+}
+
+type SubscriptionRecord = {
+  method: string
+  params: unknown
+  listener: (result: unknown) => void
+  options?: Parameters<RpcClient['subscribe']>[3]
+  disposePhysical: (() => void) | null
+  cancelled: boolean
 }
 
 type PendingRequest = {
@@ -64,13 +72,9 @@ export function createStableLogicalRpcClient(
   let generation = 1
   let closed = false
   let suspended = false
+  let nextSubscriptionId = 0
   let activeStateUnsubscribe: (() => void) | null = null
-  const subscriptions = new LogicalSubscriptionRegistry({
-    isClosed: () => closed,
-    isSuspended: () => suspended,
-    activeSession: () => activeSession,
-    generation: () => generation
-  })
+  const subscriptions = new Map<number, SubscriptionRecord>()
   const pendingRequests = new Set<PendingRequest>()
   const stateListeners = new Set<(state: ConnectionState) => void>()
   let state = initialSession.getState()
@@ -116,11 +120,44 @@ export function createStableLogicalRpcClient(
       if (closed) {
         return () => {}
       }
-      return subscriptions.add(method, params, listener, options)
+      const id = ++nextSubscriptionId
+      const record: SubscriptionRecord = {
+        method,
+        params,
+        listener,
+        options,
+        disposePhysical: null,
+        cancelled: false
+      }
+      subscriptions.set(id, record)
+      if (!suspended) {
+        attachSubscription(record, activeSession, generation)
+      }
+      return () => {
+        if (record.cancelled) {
+          return
+        }
+        record.cancelled = true
+        record.disposePhysical?.()
+        record.disposePhysical = null
+        subscriptions.delete(id)
+      }
     },
 
     updateTerminalSubscriptionViewport(terminal, viewport) {
-      subscriptions.updateTerminalViewport(terminal, viewport)
+      for (const record of subscriptions.values()) {
+        if (
+          record.params &&
+          typeof record.params === 'object' &&
+          'terminal' in record.params &&
+          record.params.terminal === terminal
+        ) {
+          record.params = { ...record.params, viewport }
+        }
+      }
+      if (!suspended) {
+        activeSession.updateTerminalSubscriptionViewport(terminal, viewport)
+      }
     },
 
     getState: () => state,
@@ -143,7 +180,10 @@ export function createStableLogicalRpcClient(
       closed = true
       activeStateUnsubscribe?.()
       activeStateUnsubscribe = null
-      subscriptions.disposeAll()
+      for (const record of subscriptions.values()) {
+        record.disposePhysical?.()
+      }
+      subscriptions.clear()
       // Why: let the physical close settle in-flight requests — it knows which
       // frames were written and marks those delivery-unknown; a blanket local
       // reject would erase that distinction.
@@ -158,7 +198,10 @@ export function createStableLogicalRpcClient(
       suspended = true
       activeStateUnsubscribe?.()
       activeStateUnsubscribe = null
-      subscriptions.detachAll()
+      for (const record of subscriptions.values()) {
+        record.disposePhysical?.()
+        record.disposePhysical = null
+      }
       // Why: let the physical close settle in-flight requests — it knows which
       // frames were written and marks those delivery-unknown (a suspend can cut
       // over a half-open relay whose sends may already be delivered).
@@ -211,7 +254,11 @@ export function createStableLogicalRpcClient(
 
       // Why: replay on the authenticated replacement before closing the old
       // session, but fence callbacks until the generation becomes current.
-      subscriptions.replayOnto(nextSession, nextGeneration)
+      for (const record of subscriptions.values()) {
+        const disposePrevious = record.disposePhysical
+        attachSubscription(record, nextSession, nextGeneration)
+        disposePrevious?.()
+      }
       generation = nextGeneration
       activeSession = nextSession
       activePath = path
@@ -256,6 +303,23 @@ export function createStableLogicalRpcClient(
     if (failed && forwarder.forwarded() && state !== 'connected') {
       publishState('disconnected')
     }
+  }
+
+  function attachSubscription(
+    record: SubscriptionRecord,
+    session: RpcClient,
+    subscriptionGeneration: number
+  ): void {
+    record.disposePhysical = session.subscribe(
+      record.method,
+      record.params,
+      (result) => {
+        if (!closed && !record.cancelled && generation === subscriptionGeneration) {
+          record.listener(result)
+        }
+      },
+      record.options
+    )
   }
 
   function bindActiveState(session: RpcClient, sessionGeneration: number): void {
